LayoutTests/fast/notifications/notifications-request-permission-optional.html

<!DOCTYPE html>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script type="text/javascript">
shouldNotBe('window.webkitNotifications', 'undefined');
shouldNotThrow('window.webkitNotifications.requestPermission()');
</script>
<script src="../js/resources/js-test-post.js"></script>
</head>
</html>
